pyproject.toml
[build-system]
requires = ["maturin>=1.7,<2.0"]
build-backend = "maturin"

[project]
name = "dim-python"
requires-python = ">=3.8"
classifiers = [
    "Programming Language :: Rust",
    "Programming Language :: Python :: Implementation :: CPython",
    "Programming Language :: Python :: Implementation :: PyPy",
]
dynamic = ["version"]
dependencies = [
    "pillow>=10.4.0",
]
[tool.maturin]
features = ["pyo3/extension-module"]

[tool.uv]
dev-dependencies = [
    "pytest>=8.3.4",
]
